requirements.txt
numpy==1.26.4
open3d==0.18.0
matplotlib==3.8.4
plotly==5.21.0
scipy==1.13.0
scikit-learn==1.4.2
umap-learn==0.5.6
torch==2.2.2
plyfile==1.0.3
plotly==5.21.0